+++ client/src/io/worxbend/gitea4s/api/ReposApi.scala
@@ -1,8 +1,16 @@
 package io.worxbend.gitea4s.api
 
 import io.worxbend.gitea4s.error.GiteaError
-import io.worxbend.gitea4s.http.RepoListParams
-import io.worxbend.gitea4s.model.{Branch, NewIssuePinsAllowed, Repository, Tag}
+import io.worxbend.gitea4s.http.{CommitStatusListParams, RepoListParams}
+import io.worxbend.gitea4s.model.{
+  Branch,
+  CombinedStatus,
+  CommitStatus,
+  CreateStatusOption,
+  NewIssuePinsAllowed,
+  Repository,
+  Tag
+}
 import zio.{Chunk, IO}
 import zio.stream.ZStream
 
@@ -18,3 +26,26 @@ trait ReposApi:
   def branches(owner: String, repo: String): ZStream[Any, GiteaError, Branch]
 
   def tags(owner: String, repo: String): ZStream[Any, GiteaError, Tag]
+
+  def combinedStatusByRef(owner: String, repo: String, ref: String): IO[GiteaError, CombinedStatus]
+
+  def statusesByRef(
+      owner: String,
+      repo: String,
+      ref: String,
+      params: CommitStatusListParams = CommitStatusListParams.default
+  ): ZStream[Any, GiteaError, CommitStatus]
+
+  def statuses(
+      owner: String,
+      repo: String,
+      sha: String,
+      params: CommitStatusListParams = CommitStatusListParams.default
+  ): ZStream[Any, GiteaError, CommitStatus]
+
+  def createStatus(
+      owner: String,
+      repo: String,
+      sha: String,
+      body: CreateStatusOption
+  ): IO[GiteaError, CommitStatus]